// backend/src/entities/enums.ts
export enum UserRole {
  ADMIN = 'admin',
  USER = 'user',
}

export enum LearningStatus {
  TODO = 'todo',
  DOING = 'doing',
  DONE = 'done',
}

export enum InterviewStatus {
  PENDING = 'pending',
  PASSED = 'passed',
  REJECTED = 'rejected',
  CANCELLED = 'cancelled',
}

export enum FinanceType {
  INCOME = 'income',
  EXPENSE = 'expense',
}

export enum ShoppingStatus {
  PENDING = 'pending',
  BOUGHT = 'bought',
}